requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.1.13
langchain-community==0.0.29
requests==2.31.0
python-dotenv==1.0.1